+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #1854 Support for analyses (pre) conditions on sample registration
+- #1855 Fix analyses results not set after auto import
 - #1853 Fix sample progress update after instrument results import
 - #1852 Fix `{client}` is wrongly stated as a variable on ID generation
 - #1850 Add valid password for portal_setup tarball import new user creation